% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harp.R
\name{hiv_tx_outcome}
\alias{hiv_tx_outcome}
\title{HARP Diaggregations}
\usage{
hiv_tx_outcome(
  previous_outcome,
  last_pill_date,
  report_end_date,
  cutoff,
  units = "days"
)
}
\arguments{
\item{previous_outcome}{column for existing/old reference outcome variable}

\item{last_pill_date}{date of last know pill for the client}

\item{report_end_date}{end of the reporting period for reference of the new outcome}

\item{cutoff}{number in days/months to calculate the cutoff date against}

\item{units}{type of date unit to calculate the cutoff date against
\itemize{
\item \code{"days"}
\item \code{"months"}
}}
}
\value{
returns a vector containing the updated hiv treatment outcomes
}
\description{
These functions help create the necessary disaggregations used for the various
HARP datasets.
}
