% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{log}
\alias{log}
\title{Info logger}
\usage{
log(msg, log_type = "info")
}
\arguments{
\item{msg}{message to be printed in the console}

\item{log_type}{type of logging to be done}
}
\value{
prints a formatted message in the console
}
\description{
These functions help provided formatted messages for better logging.
}
