% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{suppress_warnings}
\alias{suppress_warnings}
\title{Suppress specific warnings}
\usage{
suppress_warnings(.expr, .f, ...)
}
\arguments{
\item{.expr}{expression to be evaluated}

\item{.f}{regular expression string to find. if found, suppress}
}
\value{
returns the evaluated expression without the warnings/messages
}
\description{
This function suppresses specific warnings/messages in the console based on
the provided regular expression.
}
