% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Stata-like Tabulation}
\usage{
tab(data, ..., as_df = FALSE)
}
\arguments{
\item{data}{the input data.frame}

\item{...}{columns to aggregate with. This takes a tidyselect specification.}

\item{as_df}{should output be returned as a data.frame object. By default, the data is printed with formats.}
}
\value{
Returns an aggregated summary data.frame.
}
\description{
This function transforms a data.frame into an aggregate table. The main idea
was to replicate the \code{tab} command from Stata.
}
\examples{

# uni-variate tabulation
mtcars \%>\% tab(am)
tab(mtcars, am)

# multi-variate tabulation
mtcars \%>\% tab(am, cyl)
tab(mtcars, am, cyl)
}
