% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tabstat}
\alias{tabstat}
\title{Stata-like Statistics Tabulation}
\usage{
tabstat(data, ...)
}
\arguments{
\item{data}{the input data.frame}

\item{...}{columns to summarise. This takes a tidyselect specification.}
}
\value{
Returns an summarised statistics data.frame.
}
\description{
This function transforms a data.frame into a summary table indicating the \code{min},
\code{median}, \code{max}, and \code{NAs} of the columns specified.
}
\examples{

# uni-variate tabulation
mtcars \%>\% tabstat(am)
tabstat(mtcars, am)

# multi-variate tabulation
mtcars \%>\% tabstat(am, cyl)
tabstat(mtcars, am, cyl)
}
